Upgrade print_stack function to a global hexdump
print_stack funtion is now a global printing memory function with 2
parameters base and limit
