+++ include/kernel/string.h
@@ -13,6 +13,7 @@
 #define STRING_H
 
 #include <stddef.h>
+#include <stdint.h>
 
 /*
 	The string library defines some basic string/memory manipulation functions.
@@ -43,5 +44,6 @@ int		memcmp(const void *s1, const void *s2, size_t n);
 void	*memcpy(void *dst, const void *src, size_t n);
 void	*memmove(void *dst, const void *src, size_t n);
 void	*memset(void *s, int c, size_t n);
+void	hexdump(uint32_t *base, uint32_t limit);
 
 #endif
