+++ kernel/gdt.c
@@ -160,69 +160,3 @@ void print_gdt()
 		i++;
 	}
 }
-
-/* Print the stack from ebp to limit
- *
- * No args at the moment.
- * @ebp is the base pointer from the stack
- * @limit is the mumber of byte printed (one address == 16 bytes) and it's
- * hardcoded. In the future we can set it as parameter.
- *
- */
-void print_stack()
-{
-	uint32_t ebp;
-	asm inline volatile ("mov %%ebp, %0" :: "r"(ebp));
-	uint32_t limit = 4096;
-	uint32_t i = 0;
-	uint8_t line = 0;
-	struct kbd_event evt;
-
-	while (i < limit)
-	{
-		if (line == 25)
-		{
-			while (!KBD_poll());
-			KBD_getevent(&evt);
-			if (evt.type != KEY_PRESSED)
-				continue;
-			line = 0;
-		}
-		VGA_setforegroundcolor(VGA_COLOR_WHITE);
-		kprintf("0x%x : ", ebp);
-		uint32_t next = ebp + 16;
-		uint32_t tmp = ebp;
-
-		while (tmp < next)
-		{
-			if (*(char*)tmp > 32)
-			{
-				VGA_setforegroundcolor(VGA_COLOR_BLUE);
-				kprintf("%x ", *(char*)tmp);
-				VGA_setforegroundcolor(VGA_COLOR_WHITE);
-			}
-			else
-				kprintf("00 ");
-			tmp++;
-		}
-		next = ebp + 16;
-		tmp = ebp;
-		while (tmp < next)
-		{
-			if (*(char*)tmp > 32)
-			{
-				VGA_setforegroundcolor(VGA_COLOR_BLUE);
-				kprintf("%c", *(char *)tmp);
-				VGA_setforegroundcolor(VGA_DFL_COLOR);
-			}
-			else
-				kprintf(".");
-			tmp++;
-		}
-		VGA_setforegroundcolor(VGA_DFL_COLOR);
-		kprintf("\n");
-		i += 16;
-		ebp += 16;
-		line++;
-	}
-}
