+++ lib/string/mem.c
@@ -10,6 +10,10 @@
  */
 
 #include <stddef.h>
+#include <stdint.h>
+#include <kernel/vga.h>
+#include <kernel/keyboard.h>
+#include <kernel/print.h>
 
 void *memchr(const void *s, int c, size_t n) {
 	unsigned char *sc = (unsigned char *)s;
@@ -78,4 +82,69 @@ void *memset(void *s, int c, size_t n) {
 		*sc++ = c;
 	}
 	return ptr;
+}
+
+/* Hexdump to print from address to limit
+ *
+ * @base is the address to start the print from
+ * @limit is the size we want to print from @base
+ *
+ */
+void hexdump(uint32_t *base, uint32_t limit)
+{
+	uint32_t i = 0;
+	uint8_t line = 0;
+	struct kbd_event evt;
+	
+	VGA_clear();
+	KBD_flush(&evt);
+	while (i < limit)
+	{
+		if (line == VGA_HEIGHT)
+		{
+			line = 0;
+			while (!KBD_poll());
+			KBD_getevent(&evt);
+			if (evt.type != KEY_PRESSED) {
+				KBD_flush(&evt);
+				continue;
+			}
+		}
+		VGA_setforegroundcolor(VGA_COLOR_WHITE);
+		kprintf("0x%x:  ", base);
+		uint32_t *next = base + 16;
+		uint32_t *tmp = base;
+		
+		while (tmp < next)
+		{
+			if (*(uint8_t*)tmp > 32)
+			{
+				VGA_setforegroundcolor(VGA_COLOR_BLUE);
+				kprintf("%x ", *(uint8_t*)tmp);
+				VGA_setforegroundcolor(VGA_COLOR_WHITE);
+			}
+			else
+				kprintf("00 ");
+			tmp++;
+		}
+		next = base + 16;
+		tmp = base;
+		while (tmp < next)
+		{
+			if (*(uint8_t*)tmp > 32)
+			{
+				VGA_setforegroundcolor(VGA_COLOR_BLUE);
+				kprintf("%c", *(uint8_t *)tmp);
+				VGA_setforegroundcolor(VGA_DFL_COLOR);
+			}
+			else
+				kprintf(".");
+			tmp++;
+		}
+		VGA_setforegroundcolor(VGA_DFL_COLOR);
+		kprintf("\n");
+		i += 16;
+		base += 16;
+		line++;
+	} 
 }
